+++ steptable.h
@@ -15,7 +15,7 @@ public:
     ~StepTable();
 
 public slots:
-    void insertRow(qint16 row);
+    void insertRow(qint64 row);
     void insertNewRightStep(qint64 frameNum);
     void insertNewLeftStep(qint64 frameNum);
     void handleCellChanged(QTableWidgetItem* item);
@@ -27,9 +27,9 @@ private:
     void styleHeader();
     void setColumnNames();
     void addStep(qint64 frameNum, BodySide side);
-    void removeStep(qint16 row, qint16 col);
-    void sortColumn(qint16 col);
-    bool alreadyInColumn (qint16 col, qint64 frameNum);
+    void removeStep(qint64 row, qint64 col);
+    void sortColumn(qint64 col);
+    bool alreadyInColumn (qint64 col, qint64 frameNum);
 
     bool writeToCSV();
     bool readFromCSV();
